Move the log formatter
